feat: render imgs from svg, better line algorithm, yarn color projection

+++ src/cli.rs
@@ -94,8 +94,8 @@ pub struct Arguments {
     #[clap(short = 'o', long, value_parser=number_between_0_and_1, default_value_t = 0.3)]
     pub yarn_opacity: f64,
     /// Yarn width
-    #[clap(short = 'w', long, value_parser, default_value_t = 1)]
-    pub yarn_width: u32,
+    #[clap(short = 'w', long, value_parser, default_value_t = 1.)]
+    pub yarn_width: f32,
     /// How much to lighten the pixels at each line pass, low values encourage line overlap [0, 1]
     #[clap(short, long, value_parser=number_between_0_and_1, default_value_t = 0.5)]
     pub lighten_factor: f64,
